Backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
chromadb==0.5.0
sentence-transformers==2.7.0
rank-bm25==0.2.2
openai==1.30.1
pypdf==4.2.0
langchain==0.2.1
python-dotenv==1.0.1